Tweaked log4net settings and added the basic functionality of the service to iterate through a source directory (defined in app.config) to find all sub-folders in that directory and parse them into a min/max date range based on month/day.
The application will identify ALL matching folders and select the one with the smallest range (Most precise) as the source directory. I've also added an optional app key to allow the user to determine the percent of times an image is drawn from the root folder instead of a subfolder.