Tweaked code and added installer project so the service can be installed. There are some issues. To run as a local service, all files must exist on computer. Cloud files (OneDrive) cannot be downloaded by the local service account. Also, to modify the config file, the user has to be an administrator, which isn't reasonable. Will re-examine this later.
Also fixed a bug where the code was MOVING the file from the source directory to the destination directory instead of COPYING the file. Since the files are not sent to the recycling bin when deleted, this would have been very problematic. Fortunately, my source files were on OneDrive so I used the cloud recycle bin to restore them.